org add, edit, delete fixed

+++ main/webUi/page2Components/organizationManagement.py
@@ -61,6 +61,22 @@ class Organization(Page2Component):
 			})
 
 			# USer delete part ends
+
+			#image delete part
+			oldFileQuery = session.query(db.Info).filter(and_(db.Info.entity_id==formData['id'], db.Info.type==db.Info.Type.Image.value))
+			if(len(oldFileQuery.all()) != 0):
+				try:
+					oldInfoRow = oldFileQuery.one()
+					oldImageId = oldInfoRow.data
+					oldLogoRow = session.query(db.logo).filter(db.logo.id==oldImageId).one()
+					oldImageName = oldImageId + oldLogoRow.extension
+					session.delete(oldLogoRow)
+					session.delete(oldInfoRow)
+					os.remove(os.path.join (AppConfig.DbAssets, oldImageName))
+				except:
+					traceback.print_exc()
+			#image delete ends
+
 			db.Info.delete({
 				'entity_id':formData['id'],
 			})
@@ -180,34 +196,35 @@ class Organization(Page2Component):
 			}))
 
 			try:
-				readHandle = formData['image'].file
-				chunkSize = 1<<13 # 2^13
-				with db.session () as session :
-					newFile = db.logo ()
-					newFile.id = db.Entity.newUuid ()
-					newFile.fileName = formData['image'].filename
-					extension = newFile.fileName[-4:].lower ()
-					newFile.extension = extension
-					if extension not in ['.jpg','.jpeg','.png']:
-						self.jsonFailure('Incompatible Extension')
-					session.add (newFile)
+				if 'image' in formData:
+					readHandle = formData['image'].file
+					chunkSize = 1<<13 # 2^13
+					with db.session () as session :
+						newFile = db.logo ()
+						newFile.id = db.Entity.newUuid ()
+						newFile.fileName = formData['image'].filename
+						extension = newFile.fileName[-4:].lower ()
+						newFile.extension = extension
+						if extension not in ['.jpg','.jpeg','.png']:
+							self.jsonFailure('Incompatible Extension')
+						session.add (newFile)
 
-					with open (os.path.join (AppConfig.DbAssets, newFile.id + extension),'wb' ) as writeHandle:
-						while True :
-							data = readHandle.read (chunkSize)
-							writeHandle.write (data)
-							if not data :
-								break
+						with open (os.path.join (AppConfig.DbAssets, newFile.id + extension),'wb' ) as writeHandle:
+							while True :
+								data = readHandle.read (chunkSize)
+								writeHandle.write (data)
+								if not data :
+									break
+								#
 							#
 						#
-					#
-					worker.session.add(db.Info.newFromParams({
-						'id': db.Entity.newUuid(),
-						'entity_id': newOrganization.id,
-						'enumType': db.Info.Type.Image,
-						'preference': 0,
-						'data': newFile.id,
-					}))
+						worker.session.add(db.Info.newFromParams({
+							'id': db.Entity.newUuid(),
+							'entity_id': newOrganization.id,
+							'enumType': db.Info.Type.Image,
+							'preference': 0,
+							'data': newFile.id,
+						}))
 			except:
 				traceback.print_exc ()
 				return self.jsonFailure()
@@ -305,16 +322,18 @@ class Organization(Page2Component):
 							#
 
 							oldFileQuery = session.query(db.Info).filter(and_(db.Info.entity_id==formData['id'], db.Info.type==db.Info.Type.Image.value))
-							try:
-								oldInfoRow = oldFileQuery.one()
-								oldImageId = oldInfoRow.data
-								oldLogoRow = session.query(db.logo).filter(db.logo.id==oldImageId).one()
-								oldImageName = oldImageId + oldLogoRow.extension
-								session.delete(oldLogoRow)
-								session.delete(oldInfoRow)
-								os.remove(os.path.join (AppConfig.DbAssets, oldImageName))
-							except:
-								traceback.print_exc()
+							if (len(oldFileQuery.all()) != 0):
+								try:
+									oldInfoRow = oldFileQuery.one()
+									oldImageId = oldInfoRow.data
+									oldLogoRow = session.query(db.logo).filter(db.logo.id==oldImageId).one()
+									oldImageName = oldImageId + oldLogoRow.extension
+									session.delete(oldLogoRow)
+									session.delete(oldInfoRow)
+									os.remove(os.path.join (AppConfig.DbAssets, oldImageName))
+								except:
+									traceback.print_exc()
+
 							worker.session.add(db.Info.newFromParams({
 								'id': db.Entity.newUuid(),
 								'entity_id': formData['id'],
